Consider adding explicit OpenSSL library dependencies to binding.gyp
Greetings, thank you for UWS.

I was compiling UWS on Windows and discovered that my local environment was missing the above two binary dependencies during the linking stage. This manifested as missing SSL symbols during linker phase of compilation. 

I thought I would offer a PR to show what worked for me. I am not an expert with node-gyp, or Windows development so it's possible that there is some other problem with this approach. Still, it took me quite a lot of research (gyp is very poorly documented) to figure out how to convince the linker to include these libraries so I hope you'll consider accepting or adapting this patch for the next poor soul to encounter this problem.